requirements.txt
audioread==2.1.8
cffi==1.13.2
decorator==4.4.1
joblib==0.14.1
librosa==0.7.1
llvmlite==0.30.0
numba==0.46.0
numpy==1.18.0
pycparser==2.19
python-rapidjson==0.9.1
resampy==0.2.2
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
SoundFile==0.10.3.post1
umap-learn==0.3.10
